move to main pkg. dont crash for invalidily wrapped transaction, just gracefully wrap the transaction of the main transaction when comparing
